Revert "Move base subject/observer to AtData since they are widely shared."
This reverts commit e441722256455f7754fbe503a92cb9082e289f94.
