+++ AtData/CMakeLists.txt
@@ -14,9 +14,6 @@ Set(DEPENDENCIES
 
 set(SRCS
   # Add all the source files below this line. Those must have cxx for their extension.
-  AtDataObserver.cxx
-  AtDataSubject.cxx
-
   AtPadBase.cxx
   AtPad.cxx
   AtAuxPad.cxx
