+++ AtEventDisplay/AtEventTab/AtDataObserver.h
@@ -5,13 +5,8 @@
 
 namespace DataHandling {
 class Subject;
-/**
- * This is the base class for any object that may observe data from a Subject and reacto on an update.
- *
- * It can register itself with any Subject and of that subject changes this class will be informed.
- * @ingroup DataHandling
- */
-class Observer {
+
+   class Observer {
 private:
    std::set<Subject *> fSubjects;
