--- AtEventDisplay/AtEventTab/AtSubjectEventViewer.cxx
@@ -1,1 +1,0 @@
-#include "AtSubjectEventViewer.h"
